+++ packages/omniston-sdk/src/omniston/index.ts
@@ -24,3 +24,7 @@ export type { Logger } from "../logger/Logger";
 export { Omniston } from "./Omniston";
 export type { IOmnistonDependencies } from "./Omniston.types";
 export { OmnistonError } from "./OmnistonError";
+export type {
+  OmnistonErrorDetails,
+  OmnistonErrorInfo,
+} from "./OmnistonError.types";